cook(runner): surface active jobs in status
